my first javadoc (BeeUtils.unbox)
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@224 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/shared/Assert.java
@@ -125,11 +125,11 @@ public class Assert {
     }
   }
 
-  public static void isPositive(int x) {
+  public static void isPositive(double x) {
     isPositive(x, ASSERTION_FAILED + "(" + x + ") argument must be positive");
   }
 
-  public static void isPositive(int x, String msg) {
+  public static void isPositive(double x, String msg) {
     if (x <= 0) {
       throw new BeeRuntimeException(msg);
     }
